+++ changelog.md
@@ -3,6 +3,7 @@
 ## v1.0.1335 (2018-08-28)
 
 - Increase value of retry counter for CreateContainer call (#495)
+- Workflow validation (#497)
 
 ## v1.0.1334 (2018-08-27)